+++ src/Fable.Core/Fable.Core.JsInterop.fs
@@ -81,6 +81,10 @@ let importSideEffects (path: string): unit = jsNative
 /// Imports a file dynamically at runtime
 let importDynamic<'T> (path: string): JS.Promise<'T> = jsNative
 
+/// Imports a reference from an external file dynamically at runtime
+/// ATENTION: Needs fable-compiler 2.3, pass the reference directly in argument position (avoid pipes)
+let importValueDynamic (x: 'T): JS.Promise<'T> = jsNative
+
 /// Used when you need to send an F# record to a JS library accepting only plain JS objects (POJOs)
 let toPlainJsObj(o: 'T): obj = jsNative