+++ src/Fable.Transforms/AST/AST.Fable.fs
@@ -28,6 +28,7 @@ type Type =
     | GenericParam of name: string
     | ErasedUnion of genericArgs: Type list
     | DeclaredType of FSharpEntity * genericArgs: Type list
+    | AnonymousRecordType of fieldNames: string[] * genericArgs: Type list
 
     member this.Generics =
         match this with
@@ -206,7 +207,7 @@ type ValueKind =
         | NewRecord(_, kind, genArgs) ->
             match kind with
             | DeclaredRecord ent -> DeclaredType(ent, genArgs)
-            | AnonymousRecord _ -> Any
+            | AnonymousRecord fieldNames -> AnonymousRecordType(fieldNames, genArgs) 
         | NewUnion(_, _, ent, genArgs) -> DeclaredType(ent, genArgs)
         | NewErasedUnion(_, genArgs) -> ErasedUnion genArgs