+++ src/fable-publish-utils/PublishUtils.fs
@@ -5,6 +5,7 @@ open System.Text.RegularExpressions
 open Fable.Core
 open Fable.Core.JsInterop
 open Fable.Core.DynamicExtensions
+open Fable.Import
 
 module private Helpers =
     let readline: obj = importAll "readline"
@@ -294,7 +295,7 @@ module Publish =
         let json =
             projDir </> "package.json"
             |> readFile
-            |> Fable.Import.JS.JSON.parse
+            |> JS.JSON.parse
         json?version
 
     let bumpNpmVersion projDir newVersion =
@@ -362,7 +363,7 @@ module Publish =
 
     let pushNpm (projDir: string) buildAction =
         let checkPkgVersion json: string option =
-            (Fable.Import.JS.JSON.parse json)?version |> Option.ofObj
+            (JS.JSON.parse json)?version |> Option.ofObj
         let releaseVersion = loadReleaseVersion projDir
         if needsPublishing checkPkgVersion releaseVersion (projDir </> "package.json") then
             buildAction()
@@ -387,7 +388,7 @@ let pushNpm projDir buildAction =
     Publish.pushNpm projDir buildAction
 
 let getDotNetSDKVersionFromGlobalJson(): string =
-    let json = readFile "global.json" |> Fable.Import.JS.JSON.parse
+    let json = readFile "global.json" |> JS.JSON.parse
     json?sdk?version
 
 let installDotnetSdk() =