+++ src/quicktest/splitter.config.js
@@ -4,6 +4,7 @@ module.exports = {
   cli: { path: resolve("../Fable.Cli") },
   entry: resolve("QuickTest.fsproj"),
 //   outDir: resolve("temp"),
+//   port: 61225,
   fable: {
     define: ["DEBUG"]
   },